+++ mod.ts
@@ -7,9 +7,10 @@ export {
   _unstable_useCloudflareWorkersMode,
   type FullSoakMiddleware,
   type OakRoutingControllerClass,
+  useFetchMode,
   useFullSoak,
   type UseFullSoakOptions,
 } from "./src/useFullSoak.ts";
 export { getOrigin } from "./src/metastore.ts";
 export { setupDefaultFullsoakLogger } from "./src/utils.ts";
-export { HtmlShell } from "./src/HtmlShell.tsx";
+export { HtmlShell } from "./src/HtmlShell.ts";