+++ src/getClientSideJsForRoute.ts
@@ -3,11 +3,12 @@ import { getGlobalComponentsDirName } from "./metastore.ts";
 
 /**
  * retrieve the initial javascript to be loaded into the html document generated with
- * {@link HtmlShell} via the "static" path `/js/:compName/mount.js`; if this javascript
+ * {@link HtmlShell} via the "static" path `/components/:compName/mount`; if this javascript
  * content fails to load (for any reason), the SSR document should still be enough for
  * human or bot users (we'll lose any interactivity that's `hydrated` by this javascript though)
  */
 export const getClientSideJsForRoute = (compName: string): string => {
+  // @NOTE importing from `${compName}.tsx` is also supported as a fallback mechanism
   const importByRoute = `
   import { ${compName} as View } from "/${getGlobalComponentsDirName()}/${compName}/index.tsx";
   `;